requirements.txt
fastapi
uvicorn
torch
facenet-pytorch
opencv-python
python-multipart
Pillow
python-multipart 
pillow 
numpy

